#8 feature: contoller에 실제 Service 적용
- 사용되는 Response / Request Dto 이름 수정 및 Record와 Time 분리
- 여러 건 조회 Service 생성
- RecordTimeService에 Delete 기능 생성

+++ monitoring/src/main/java/monitoring/service/RecordTimeService.java
@@ -3,11 +3,14 @@ package monitoring.service;
 import lombok.RequiredArgsConstructor;
 import monitoring.domain.Record;
 import monitoring.domain.TimeLog;
-import monitoring.service.dto.UpdateRecordResponseDto;
+import monitoring.service.request.RecordSearchCond;
+import monitoring.service.response.UpdateRecordResult;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.StringUtils;
 
 import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -37,20 +40,43 @@ public class RecordTimeService {
         return record;
     }
 
+    public List<Record> getRecords(RecordSearchCond cond) {
+
+        List<TimeLog> timeLogs = timeLogService.searchOverlappingTimeLogs(cond.getDate(), cond.getStartTime(), cond.getEndTime());
+        Set<Record> searchRecordSet = timeLogs.stream()
+                .map(t -> this.getRecordWithTimeLogs(t.getRecord().getId()))
+                .filter(record -> {
+                    if (!StringUtils.hasText(cond.getAction())) {
+                        return true;
+                    }
+                    return record.getAction()
+                            .toLowerCase()
+                            .contains(cond.getAction().toLowerCase());
+                })
+                .collect(Collectors.toSet());
+
+        return searchRecordSet.stream().toList();
+    }
+
+
     @Transactional
     public Record update(Long id, Record recordData) {
         Record updatedRecord = recordService.update(id, recordData);
 
-        timeLogService.deleteByRecordId(id);
-        List<TimeLog> savedTimeLogs = timeLogService.create(updatedRecord, recordData.getTimeLogs());
+        if (recordData.getTimeLogs() != null && !recordData.getTimeLogs().isEmpty()) {
+            timeLogService.deleteByRecordId(id);
+            List<TimeLog> savedTimeLogs = timeLogService.create(updatedRecord, recordData.getTimeLogs());
+            connectRecordAndTimeLogList(updatedRecord, savedTimeLogs);
+        } else {
+            List<TimeLog> notUpdatedTimeLogs = timeLogService.getByRecordId(id);
+            connectRecordAndTimeLogList(updatedRecord, notUpdatedTimeLogs);
+        }
 
-        connectRecordAndTimeLogList(updatedRecord, savedTimeLogs);
         return updatedRecord;
     }
 
     @Transactional
-    public UpdateRecordResponseDto updateWithSideEffect(Long id, Record recordData) {
-
+    public UpdateRecordResult updateWithSideEffect(Long id, Record recordData) {
 
         Set<Record> deletedRecordSet = new HashSet<>();
         Set<Record> changedRecordSet = new HashSet<>();
@@ -79,9 +105,17 @@ public class RecordTimeService {
         });
 
         Record updatedRecord = update(id, recordData);
-        return new UpdateRecordResponseDto(deletedRecordSet.stream().toList(), changedRecordSet.stream().toList(), updatedRecord);
+        return new UpdateRecordResult(deletedRecordSet.stream().toList(), changedRecordSet.stream().toList(), updatedRecord);
     }
 
+    @Transactional
+    public Record delete(Long id) {
+        List<TimeLog> deletedTimeLogs = timeLogService.deleteByRecordId(id);
+        Record deleteRecord = recordService.delete(id);
+
+        deleteRecord.setTimeLogs(deletedTimeLogs);
+        return deleteRecord;
+    }
 
     private void connectRecordAndTimeLogList(Record record, List<TimeLog> timeLogs) {
         record.setTimeLogs(new ArrayList<>(timeLogs));
@@ -99,7 +133,6 @@ public class RecordTimeService {
             deletedTimeLog.setRecord(null);
         });
     }
-
 
 }
